[TEST] Add tests to account CRUD operations

+++ e-budget-be/src/main/java/com/ebudget/account/service/AccountService.java
@@ -109,6 +109,12 @@ public class AccountService implements IAccountService {
             throw new InvalidParameterException();
         }
 
+        Account account = accountRepository.findById(accountId);
+
+        if(account == null) {
+            throw new AccountNotFoundException();
+        }
+
         accountRepository.deleteById(accountId);
     }
 }
